+++ changelog.md
@@ -2,3 +2,6 @@
 
 ## 0.1.0
 - Initial release
+
+## not released
+- rename `ToPgpWordAccumulation` in `PgpWordAggregationBitConverter` to `ToPgpWordAggregation`